apps/backend-graphene/users/tests/factories.py
import factory
from django.contrib.auth import get_user_model
from django.contrib.auth.hashers import make_password

User = get_user_model()


class UserFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = User

    email = factory.Sequence(lambda n: f"user{n}@example.com")
    username = factory.Sequence(lambda n: f"user_{n:08d}")
    password = factory.LazyFunction(lambda: make_password("testpass123!"))
    is_active = True
    email_verified = False
