apps/backend-graphene/pyproject.toml
[project]
name = "backend-graphene"
version = "0.1.0"
description = "TaskFlow backend with Django + Graphene"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "graphene-django>=3.2.3",
    "graphql-core-promise>=3.4.2",
    "pillow>=12.1.1",
    "promise>=2.3",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "pytest-factoryboy>=2.8.1",
    "ruff>=0.15.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["N805"]
